pyproject.toml
[project]
name = "enterprise-support-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.100.0",
    "httpx[http2]>=0.28.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "lxml>=6.0.1",
    "markdownify>=1.2.0",
    "numpy>=2.3.3",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "python-dotenv>=1.0.0",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.5",
    "ruff>=0.13.1",
    "sentence-transformers>=3",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "transformers>=4.44",
    "uvicorn>=0.20.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["F", "E", "I"]   # errors + import sort
